Change function to edit/populate
